authorization to posts

+++ app/graphql/mutations/delete_post.rb
@@ -8,8 +8,15 @@ class Mutations::DeletePost < Mutations::BaseMutation
     field :errors, [String], null: false
 
     def resolve(args)
+        unless context[:current_user]
+            raise Exception, "Sign in to do this action"
+        end
     
         post = Post.find(args[:id])
+
+        unless context[:current_user] == post.user 
+            raise Exception, "You do not have permission"
+        end
 
         if post.destroy 
             return {
@@ -27,6 +34,9 @@ class Mutations::DeletePost < Mutations::BaseMutation
 
         rescue ActiveRecord::RecordInvalid => invalid
             return { errors: invalid.record.errors.full_messages, deleted: false }
+
+        rescue Exception => e
+            { errors: e.message.split(",") , deleted: false}
     end
 
 end
